refactor: 더 명확한 이름으로 컴포넌트 네이밍 변경
